matrix-indication-list/requirements.txt
ipython>=8.10
jupyterlab>=3.0
kedro-datasets>=3.0
kedro-viz>=6.7.0
kedro[jupyter]~=0.19.10
notebook
scikit-learn~=1.5.1
setuptools; python_version >= "3.12"
pandas
pyspark
openpyxl
google.generativeai
vertexai
rdflib
matplotlib
asyncio
aiohttp
openai
ontobio
networkx
hdfs
s3fs
matplotlib